add buildactivity,(build.*activity)

+++ app/src/main/java/com/superbgoal/caritasrig/data/model/InternalHardDrive.kt
@@ -3,7 +3,7 @@ package com.superbgoal.caritasrig.data.model
 data class InternalHardDrive(
     val name: String,
     val price: Double,
-    val capacity: Int, // Capacity in GB
+    val capacity: Double, // Capacity in GB
     val pricePerGb: Double,
     val type: String,
     val cache: Int, // Cache size in MB
